[example] Simplify ssr examples
