+++ examples/nextjs-next/pages/_app.js
@@ -1,16 +1,11 @@
 import React from 'react';
 import App, { Container } from 'next/app';
 import Head from 'next/head';
-import { StylesProvider, ThemeProvider } from '@material-ui/styles';
+import { ThemeProvider } from '@material-ui/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
-import getPageContext from '../src/getPageContext';
+import theme from '../src/theme';
 
 class MyApp extends App {
-  constructor() {
-    super();
-    this.pageContext = getPageContext();
-  }
-
   componentDidMount() {
     // Remove the server-side injected CSS.
     const jssStyles = document.querySelector('#jss-server-side');
@@ -21,27 +16,19 @@ class MyApp extends App {
 
   render() {
     const { Component, pageProps } = this.props;
+
     return (
       <Container>
         <Head>
           <title>My page</title>
         </Head>
-        {/* Wrap every page in Styles and Theme providers */}
-        <StylesProvider
-          generateClassName={this.pageContext.generateClassName}
-          sheetsRegistry={this.pageContext.sheetsRegistry}
-          sheetsManager={this.pageContext.sheetsManager}
-        >
-          {/* ThemeProvider makes the theme available down the React
-              tree thanks to React context. */}
-          <ThemeProvider theme={this.pageContext.theme}>
-            {/* CssBaseline kickstart an elegant, consistent, and simple baseline to build upon. */}
-            <CssBaseline />
-            {/* Pass pageContext to the _document though the renderPage enhancer
-                to render collected styles on server-side. */}
-            <Component pageContext={this.pageContext} {...pageProps} />
-          </ThemeProvider>
-        </StylesProvider>
+        {/* ThemeProvider makes the theme available down the React
+            tree thanks to React context. */}
+        <ThemeProvider theme={theme}>
+          {/* CssBaseline kickstart an elegant, consistent, and simple baseline to build upon. */}
+          <CssBaseline />
+          <Component {...pageProps} />
+        </ThemeProvider>
       </Container>
     );
   }
