+++ examples/ssr-next/server.js
@@ -1,9 +1,8 @@
 import express from 'express';
 import React from 'react';
 import ReactDOMServer from 'react-dom/server';
-import { SheetsRegistry } from 'jss';
 import { createMuiTheme } from '@material-ui/core/styles';
-import { StylesProvider, ThemeProvider, createGenerateClassName } from '@material-ui/styles';
+import { ServerStyleSheets, ThemeProvider } from '@material-ui/styles';
 import green from '@material-ui/core/colors/green';
 import red from '@material-ui/core/colors/red';
 import App from './App';
@@ -33,28 +32,19 @@ const theme = createMuiTheme({
 });
 
 function handleRender(req, res) {
-  // Create a sheetsRegistry instance.
-  const sheetsRegistry = new SheetsRegistry();
-  // Create a sheetsManager instance.
-  const sheetsManager = new Map();
-  // Create a new class name generator.
-  const generateClassName = createGenerateClassName();
+  const sheets = new ServerStyleSheets();
 
   // Render the component to a string.
   const html = ReactDOMServer.renderToString(
-    <StylesProvider
-      generateClassName={generateClassName}
-      sheetsRegistry={sheetsRegistry}
-      sheetsManager={sheetsManager}
-    >
+    sheets.collect(
       <ThemeProvider theme={theme}>
         <App />
-      </ThemeProvider>
-    </StylesProvider>,
+      </ThemeProvider>,
+    ),
   );
 
-  // Grab the CSS from our sheetsRegistry.
-  const css = sheetsRegistry.toString();
+  // Grab the CSS from our sheets.
+  const css = sheets.toString();
 
   // Send the rendered page back to the client.
   res.send(renderFullPage(html, css));
